requirements.txt
anthropic>=0.20.0
python-dotenv>=1.0.0
requests>=2.30.0
